+++ alembic/versions/c3d4e5f6a7b8_add_user_scopes.py
@@ -1,0 +1,28 @@
+"""add user scopes
+
+Revision ID: c3d4e5f6a7b8
+Revises: b7c1d2e3f4a5
+Create Date: 2026-06-12 13:30:00.000000
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'c3d4e5f6a7b8'
+down_revision: Union[str, Sequence[str], None] = 'b7c1d2e3f4a5'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    with op.batch_alter_table('users', schema=None) as batch_op:
+        batch_op.add_column(sa.Column('scopes', sa.Text(), nullable=True))
+
+
+def downgrade() -> None:
+    with op.batch_alter_table('users', schema=None) as batch_op:
+        batch_op.drop_column('scopes')